requirements.txt
asn1crypto==1.0.1
astroid==2.3.1
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.6
djangorestframework==3.10.3
fabric==2.5.0
fake==0.8
Faker==2.0.2
get==2019.4.13
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.8
invoke==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
paramiko==2.6.0
post==2019.4.13
public==2019.4.13
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.2
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
urllib3==1.25.6
websocket-client==0.48.0
wrapt==1.11.2
